# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Signal Analyzer API - All 10 lecture points
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signal/transform:
    post:
      operationId: transformSignal
      tags: [signal]
      summary: "Point 1 - Signal Transformation"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransformRequest"
      responses:
        "200":
          description: Transformation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransformResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /signal/ztransform:
    post:
      operationId: computeZTransform
      tags: [signal]
      summary: "Point 9 - Z-Transform and DFT"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZTransformRequest"
      responses:
        "200":
          description: Z-Transform and DFT result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZTransformResult"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TransformRequest:
      type: object
      properties:
        tValues:
          type: array
          items:
            type: number
        xValues:
          type: array
          items:
            type: number
        expression:
          type: string
        domain:
          type: string
          enum: [continuous, discrete]
      required: [tValues, xValues, expression, domain]
    TransformStep:
      type: object
      properties:
        oldT:
          type: number
        expression:
          type: string
        newT:
          type: number
        xValue:
          type: number
      required: [oldT, expression, newT, xValue]
    TransformResult:
      type: object
      properties:
        latexFormula:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TransformStep"
        newTValues:
          type: array
          items:
            type: number
        xValues:
          type: array
          items:
            type: number
        summary:
          type: string
      required: [latexFormula, steps, newTValues, xValues, summary]

    # Point 2 - Odd/Even
...
    ZTransformRequest:
      type: object
      properties:
        xValues:
          type: array
          items:
            type: number
        nStart:
          type: number
          description: Starting index n
        domain:
          type: string
          enum: [discrete]
        computeDFT:
          type: boolean
        computeZTransform:
          type: boolean
      required: [xValues, nStart, domain]
    ZTransformStep:
      type: object
      properties:
        n:
          type: number
        xN:
          type: number
        term:
          type: string
          description: x[n]*z^(-n) term
      required: [n, xN, term]
    DFTBin:
      type: object
      properties:
        k:
          type: number
        magnitude:
          type: number
        phase:
          type: number
        realPart:
          type: number
        imagPart:
          type: number
        formulaDetail:
          type: string
      required: [k, magnitude, phase, realPart, imagPart, formulaDetail]
    ZTransformResult:
      type: object
      properties:
        latexZTransform:
          type: string
        latexDFT:
          type: string
        zTransformTerms:
          type: array
          items:
            $ref: "#/components/schemas/ZTransformStep"
        zTransformPolynomial:
          type: string
        dftBins:
          type: array
          items:
            $ref: "#/components/schemas/DFTBin"
        rocDescription:
          type: string
        summary:
          type: string
      required: [latexZTransform, latexDFT, zTransformTerms, zTransformPolynomial, dftBins, rocDescription, summary]

    # Point 10 - Final Report
...